+++ src/wenet_service_api/api/ws/ws.py
@@ -30,7 +30,6 @@ class WsInterface:
             (WeNetUserProfileInterfaceBuilder.routes(service_connector_collector, self._authorized_api_key), "/user"),
             (TaskResourceInterfaceBuilder.routes(service_connector_collector, self._authorized_api_key), "/task"),
             (TaskTransactionInterfaceBuilder.routes(service_connector_collector, self._authorized_api_key), "/task/transaction"),
-            #(MessageInterfaceBuilder.routes(self._authorized_api_key, self._dao_collector), "/messages"),
             (AppResourceInterfaceBuilder.routes(service_connector_collector, self._authorized_api_key), "/app"),
             (TaskListResourceInterfaceBuilder.routes(service_connector_collector, self._authorized_api_key), "/tasks"),
             (TokenDetailsInterfaceBuilder.routes(service_connector_collector, self._authorized_api_key), "/token"),