+++ test/unit/wenet_service_api/api/ws/resources/test_app_interface.py
@@ -5,8 +5,9 @@ from datetime import datetime
 
 from mock import Mock
 
-from tests.wenet_service_api_test.api.common.common_test_case import CommonTestCase
 from wenet.model.app import AppDTO, App, AppStatus
+
+from test.unit.wenet_service_api.api.common.common_test_case import CommonTestCase
 from wenet_service_api.api.ws.resource.common import WenetSource
 from wenet_service_api.common.exception.exceptions import ResourceNotFound