+++ test/unit/wenet_service_api/api/ws/resources/test_task_list_interface.py
@@ -4,9 +4,10 @@ import json
 
 from mock import Mock
 
-from tests.wenet_service_api_test.api.common.common_test_case import CommonTestCase
 from wenet.model.norm import Norm, NormOperator
 from wenet.model.task.task import TaskPage, Task, TaskGoal
+
+from test.unit.wenet_service_api.api.common.common_test_case import CommonTestCase
 from wenet_service_api.api.ws.resource.common import WenetSource